+++ src/commands/clientPackageCommand.ts
@@ -7,7 +7,6 @@ import { ILogger } from "unitejs-framework/dist/interfaces/ILogger";
 import { IncludeMode } from "../configuration/models/unite/includeMode";
 import { ScriptIncludeMode } from "../configuration/models/unite/scriptIncludeMode";
 import { UniteClientPackage } from "../configuration/models/unite/uniteClientPackage";
-import { UniteClientPackageTranspile } from "../configuration/models/unite/uniteClientPackageTranspile";
 import { UniteConfiguration } from "../configuration/models/unite/uniteConfiguration";
 import { EngineCommandBase } from "../engine/engineCommandBase";
 import { EngineVariables } from "../engine/engineVariables";
@@ -94,24 +93,11 @@ export class ClientPackageCommand extends EngineCommandBase implements IEngineCo
         clientPackage.isPackage = args.isPackage || clientPackage.isPackage;
         clientPackage.noScript = args.noScript || clientPackage.noScript;
         clientPackage.assets = args.assets || clientPackage.assets;
-        if (args.transpileAlias || (clientPackage.transpile && clientPackage.transpile.alias)) {
-            if (!clientPackage.transpile) {
-                clientPackage.transpile = new UniteClientPackageTranspile();
-            }
-            clientPackage.transpile.alias = args.transpileAlias || clientPackage.transpile.alias;
-            clientPackage.transpile.language = args.transpileLanguage || clientPackage.transpile.language;
-            clientPackage.transpile.sources = args.transpileSources || clientPackage.transpile.sources;
-            clientPackage.transpile.modules = args.transpileModules || clientPackage.transpile.modules;
-            clientPackage.transpile.stripExt = args.transpileStripExt || clientPackage.transpile.stripExt;
-        }
 
         try {
             clientPackage.testingAdditions = this.mapParser(args.testingAdditions) || clientPackage.testingAdditions;
             clientPackage.map = this.mapParser(args.map) || clientPackage.map;
             clientPackage.loaders = this.mapParser(args.loaders) || clientPackage.loaders;
-            if (clientPackage.transpile) {
-                clientPackage.transpile.transforms = this.mapFromArrayParser(args.transpileTransforms) || clientPackage.transpile.transforms;
-            }
         } catch (err) {
             this._logger.error("Input failure", err);
             return 1;
@@ -181,26 +167,6 @@ export class ClientPackageCommand extends EngineCommandBase implements IEngineCo
         }
         if (clientPackage.noScript) {
             this._logger.info("noScript", { noScript: clientPackage.noScript });
-        }
-        if (clientPackage.transpile) {
-            if (clientPackage.transpile.alias) {
-                this._logger.info("transpileAlias", { transpileAlias: clientPackage.transpile.alias });
-            }
-            if (clientPackage.transpile.language) {
-                this._logger.info("transpileLanguage", { transpileLanguage: clientPackage.transpile.language });
-            }
-            if (clientPackage.transpile.sources) {
-                this._logger.info("transpileSources", { transpileSrc: clientPackage.transpile.sources });
-            }
-            if (clientPackage.transpile.modules) {
-                this._logger.info("transpileModules", { transpileSrc: clientPackage.transpile.modules });
-            }
-            if (clientPackage.transpile.stripExt !== undefined) {
-                this._logger.info("transpileStripExt", { transpileStripExt: clientPackage.transpile.stripExt });
-            }
-            if (clientPackage.transpile.transforms) {
-                this._logger.info("transpileTransforms", { transpileSrc: clientPackage.transpile.transforms });
-            }
         }
 
         this._logger.info("");